Revert "configPath cleanup"
This reverts commit 9d2ae4ea64b6e2bad2f96e92cfcb273b0f4d0751.
